Refactor transaction-entry components for improved structure and functionality
- Remove deprecated `TransactionEntryList` and related exports from the common transaction-entry module.
- Introduce new `Data`, `TransactionEntryList`, and `Item` components with enhanced data handling and rendering logic.
- Implement loading states with `Pending` components for better user experience during data fetching.
- Organize transaction entry types (Text, Gallery, Location, Package, Personal) into a structured format for clarity and maintainability.
- Update import paths in buyer and seller transaction screens to reflect the new component structure.
