--- apps/app/src/app/@common/transaction-entry/index.ts
@@ -1,1 +1,0 @@
-export * from "./TransactionEntryList";
