+++ apps/app/src/app/@common/transaction-entry/ui/TransactionEntryList/Item/Data.tsx
