+++ apps/app/src/app/@common/transaction-entry/ui/TransactionEntryList/TransactionEntryList.tsx
