+++ apps/app/src/app/@common/transaction-entry/ui/TransactionEntryList/index.ts
@@ -1,0 +1,1 @@
+export { TransactionEntryList } from "./TransactionEntryList";
